Move messages pkg at server root
